// src/services/types/platform/hands_on_demo.ts
export type HandsOnDemoApiResponse = {
  _id: string
  __v: number
  hands_on_demo: {
    section1: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
    }
    section2: {
      image1: string
      image2: string
      image3: string
      image4: string
      image5: string
      image6: string
    }
    section3: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
    }
    section4: {
      title1_en: string
      title1_fa: string
      button1_en: string
      button1_fa: string
    }
    section5: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
    }
  }
}